+++ js/return-behavior.js
@@ -1,4 +1,4 @@
-import changeScreen from './change-screen';
+import changeScreen from './render-screen';
 import renderBack from './screen-greeting';
 
 export default (screen) => {
